+++ upgrade/upg.rb
@@ -13,27 +13,8 @@ def pre_upgrade
     system("sudo rm -rf /home/sm")
 end
 
-def os
-    @os ||= (
-        host_os = RbConfig::CONFIG['host_os']
-        shared_gh_url = "https://raw.githubusercontent.com/secman-team/install/HEAD/install"
-        lin = "linux".yellow
-        osx = "osx".black
-
-        case host_os
-        when /mswin|msys|mingw|cygwin|bccwin|wince|emc/
-            :windows
-            puts "secman upg command is only supported for #{lin} and #{osx}"
-        when /darwin|mac os/
-            :macosx
-            system("curl -fsSL #{shared_gh_url}_osx.sh | bash")
-        when /linux/
-            :linux
-            system("curl -fsSL #{shared_gh_url}_linux.sh | bash")
-        else
-            raise Error::WebDriverError, "unknown os: #{host_os.inspect}"
-        end
-    )
+def core
+    system("curl -fsSL https://secman-team.github.io/install/install.sh | bash")
 end
 
 sm = "secman".cyan
@@ -44,7 +25,7 @@ if $l == $c
     
 elsif $l != $c
     pre_upgrade()
-    os()
+    core()
 
-    puts "#{sm} was upgraded successfully"
+    puts "#{sm} was upgraded successfully 🎊"
 end