+++ CHANGELOG.md
@@ -4,7 +4,21 @@
 
 ---
 
-## [[v5.3.0] 2021-03-10](#v521-2021-03-10)
+## [[v5.3.1] 2021-03-26](#v531-2021-03-26)
+
+### Added
+
+- Show warn message when secman dependencies are not found
+
+### Changed
+
+- Remove Dockerfile
+
+### Fixed
+
+nothing
+
+## [[v5.3.0] 2021-03-10](#v530-2021-03-10)
 
 ### Added